refactor(backend): split requests.ts + admin/services.ts into submodules
Two backend route monoliths decomposed by concern. Route URLs, payloads,
status codes, error strings, logging + notification side-effects unchanged.

packages/backend/src/routes/requests.ts (545 → 17 lines)
  requests/list.ts         — GET / + /stats + /:id/resolve
  requests/create.ts       — POST / + /collection
  requests/lifecycle.ts    — approve / decline / update / delete
  requests/maintenance.ts  — search-missing + cleanup

packages/backend/src/routes/admin/services.ts (378 → 15 lines)
  services/crud.ts      — service CRUD + test + schemas
  services/helpers.ts   — Plex token + *arr profiles/rootfolders
  services/webhooks.ts  — per-service webhook status/enable/disable

Orchestrator `register(subRoute)` inherits the outer prefix — same pattern
used for the TMDB split (PR #155). Adding a new handler in either area is
now a one-file edit instead of hunting through hundreds of lines.

Follow-up to the god-file audit triggered by #138 (frontend refactor
landed in PR #157). Backend leftovers identified by the audit are now
clean too.
